Reorganize src/ into plugins/ directory, remove POC plugins and demos
Split monolithic plugins.ts into individual files under src/plugins/.
Move postgres, fiber, error plugins into src/plugins/ with updated imports.
Remove generic POC plugins (db, kv, api, jwt, crypto) — replaced by
real-world plugins that mirror actual libraries (see #7).
Delete demo files (replaced by tests in #3).
Add src/index.ts as public API entry point.
Update VISION.md §5 and §7 to reflect new structure.

Closes #2
