+++ src/index.ts
@@ -1,0 +1,17 @@
+// Public API
+export { ilo } from "./core";
+export type { Expr, ASTNode, Program, PluginDefinition, PluginContext, Plugin } from "./core";
+
+// Structural plugins
+export { num } from "./plugins/num";
+export type { NumMethods } from "./plugins/num";
+export { str } from "./plugins/str";
+export type { StrMethods } from "./plugins/str";
+
+// Real-world plugins
+export { postgres } from "./plugins/postgres";
+export type { PostgresMethods, PostgresConfig } from "./plugins/postgres";
+export { fiber } from "./plugins/fiber";
+export type { FiberMethods } from "./plugins/fiber";
+export { error } from "./plugins/error";
+export type { ErrorMethods } from "./plugins/error";
