+++ src/plugins/error.ts
@@ -31,7 +31,7 @@
 //
 // ============================================================
 
-import type { PluginDefinition, PluginContext, Expr, ASTNode } from "./core";
+import type { PluginDefinition, PluginContext, Expr, ASTNode } from "../core";
 
 // ---- What the plugin adds to $ ----------------------------
